CREXDATA System Components/NeRF Learning and Visualization (TUC)/requirements.txt
ninja
trimesh
cmake
splines
numpy==1.26.4
tqdm==4.67.1
imageio==2.37.0
nerfacc==0.5.3
nerfview==0.1.3
confluent-kafka==2.12.2
jupyterlab>=4
pandas==2.2.3
einops==0.8.1
lpips==0.1.4
jaxtyping==0.3.2
opencv-python
pytorch-msssim
tensorboard
viser
rich
matplotlib

